[ALPS07231246] s0 mp1 coverity fix
s0 mp1 coverity fix

MTK-Commit-Id: c52ac35c5fc1d8f16222cb88d008bc9e3bc3a3cf

CR-Id: ALPS07231246
Feature: Wi-Fi Driver

+++ mgmt/p2p_role_fsm.c
@@ -4419,7 +4419,7 @@ static void trimAcsScanList(IN struct ADAPTER *prAdapter,
 #if (CFG_SUPPORT_AVOID_DESENSE == 1)
 		if (IS_CHANNEL_IN_DESENSE_RANGE(prAdapter,
 			prRfChannelInfo1->ucChannelNum,
-			prRfChannelInfo1->eBand))
+			(uint32_t)prRfChannelInfo1->eBand))
 			continue;
 #endif
 		DBGLOG(P2P, INFO, "acs trim scan list, [%d]=%d %d\n",
